Remove usage of font
Loading a nonexistent front caused the error. 
This PR is purely to fox the error as I dont know what you are expecting the image to look

I removed some useless parameters in the `draw.text()` function and fixed the xy parameter format (should be in tuple: `(x, y)`)

Now it doesn't throw an error and actually sends an image. You can set the position of text, etc